more abstraction to BaseSwerveSubsystem
