+++ src/main/java/frc/robot/subsystems/swerve/SwerveSubsystem.java
@@ -15,7 +15,7 @@ import edu.wpi.first.wpilibj.Timer;
 
 import static frc.robot.Constants.SwerveConstants.*;
 
-public class SwerveSubsystem extends AbstractSwerveSubsystem {
+public class SwerveSubsystem extends BaseSwerveSubsystem {
     private final SwerveModule.TopLeft topLeftModule;
     private final SwerveModule.TopRight topRightModule;
     private final SwerveModule.BottomLeft bottomLeftModule;
@@ -23,39 +23,24 @@ public class SwerveSubsystem extends AbstractSwerveSubsystem {
 
     private final SwerveDrivePoseEstimator poseEstimator;
 
-    private final SwerveDriveKinematics kinematics;
-
     public static final double MAX_VEL = Units.feetToMeters(16.10); // Max robot tangential velocity, in m/s
     public static final double MAX_ACCEL = 3; // Max robot tangential acceleration, in m/s^2
     public static final double MAX_OMEGA = MAX_VEL / tlPos.getNorm(); // Max robot angular velocity, in rads/s (omega =
                                                                       // v / r)
 
-    private final Timer lockTimer;
-    private static final double LOCK_TIMEOUT_SECONDS = 1.0; // The elapsed idle time to wait before locking
-    private static final boolean LOCKING_ENABLE = true;
 
-    // The driver or auton commanded `SwerveModuleState` setpoints for each module;
-    // states are given in a tuple of [top left, top right, bottom left, bottom
-    // right].
-    private SwerveModuleState[] states = {
-            new SwerveModuleState(),
-            new SwerveModuleState(),
-            new SwerveModuleState(),
-            new SwerveModuleState()
-    };
+    
 
     public SwerveSubsystem() {
+        super(MAX_VEL, MAX_ACCEL, MAX_OMEGA, new SwerveDriveKinematics(
+            tlPos, trPos, blPos, brPos));
+        
         // Initialize swerve modules
         topLeftModule = new SwerveModule.TopLeft(tlDrive, tlSteer, tlOffsetRads);
         topRightModule = new SwerveModule.TopRight(trDrive, trSteer, trOffsetRads);
         bottomLeftModule = new SwerveModule.BottomLeft(blDrive, blSteer, blOffsetRads);
         bottomRightModule = new SwerveModule.BottomRight(brDrive, brSteer, brOffsetRads);
 
-        // Initialize system kinematics with top left, top right, bottom left, and
-        // bottom right swerve
-        // module positions.
-        kinematics = new SwerveDriveKinematics(
-                tlPos, trPos, blPos, brPos);
 
         // Initialize NaxX and pose estimator
 
@@ -68,60 +53,11 @@ public class SwerveSubsystem extends AbstractSwerveSubsystem {
                 new MatBuilder<>(Nat.N3(), Nat.N1()).fill(0.02, 0.02, 0.01),
                 // Vision measurement standard deviations: [X, Y, theta]
                 new MatBuilder<>(Nat.N3(), Nat.N1()).fill(0.1, 0.1, 0.01));
-
-        lockTimer = new Timer();
     }
 
-    /**
-     * Sets the swerve module states of this subsystem from provided field-centric
-     * swerve drive powers.
-     * 
-     * @param xPower       The power [-1.0, 1.0] in the x (forward) direction.
-     * @param yPower       The power [-1.0, 1.0] in the y (left) direction.
-     * @param angularPower The angular (rotational) power [-1.0, 1.0].
-     */
-    public void setSwerveDrivePowers(double xPower, double yPower, double angularPower) {
-        // If drivers are sending no input, stop all modules but hold their current
-        // angle.
-        if (xPower == 0.0 && yPower == 0.0 && angularPower == 0.0) {
-            this.states[0] = new SwerveModuleState(0.0, this.states[0].angle);
-            this.states[1] = new SwerveModuleState(0.0, this.states[1].angle);
-            this.states[2] = new SwerveModuleState(0.0, this.states[2].angle);
-            this.states[3] = new SwerveModuleState(0.0, this.states[3].angle);
-            return;
-        }
+    
 
-        // Scale [-1.0, 1.0] powers to desired velocity, turning field-relative powers
-        // into robot relative chassis speeds.
-        ChassisSpeeds speeds = ChassisSpeeds.fromFieldRelativeSpeeds(
-                xPower * MAX_VEL,
-                yPower * MAX_VEL,
-                angularPower * MAX_OMEGA,
-                super.getGyroForField());
-
-        // Calculate swerve module states from desired chassis speeds
-        this.states = kinematics.toSwerveModuleStates(speeds);
-    }
-
-    /**
-     * Sets the swerve module states of this subsystem. Module states are assumed to
-     * be passed
-     * in a tuple of [top left, top right, bottom left, bottom right].
-     * 
-     * @param states The swerve module states to set.
-     */
-    public void setSwerveModuleStates(SwerveModuleState... states) {
-        this.states = states;
-    }
-
-    /**
-     * Gets the subsystems `SwerveDriveKinematics` instance.
-     * 
-     * @return The SwerveDriveKinematics representing this system's kinematics.
-     */
-    public SwerveDriveKinematics getKinematics() {
-        return kinematics;
-    }
+   
 
     @Override
     public void periodic() {
@@ -210,11 +146,5 @@ public class SwerveSubsystem extends AbstractSwerveSubsystem {
         resetPosition(new Pose2d());
     }
 
-    /**
-     * Gets the gyro angle given by the NavX AHRS, inverted to be counterclockwise
-     * positive.
-     * 
-     * @return The robot heading as a Rotation2d.
-     */
 
 }
